refactor(data-store): extract sqlite-output-spool from sqlite-adapter
Moves the file-backed output spool helpers out of sqlite-adapter into sqlite-output-spool, behavior unchanged.
One atomic slice of the step-4 large-file decomposition.
